requirements.txt
dlib==18.17.100
nvidia-ml-py3==7.352.0
opencv-python==3.4.1.15
Pillow==5.4.1
torch==1.4.0
torchvision==0.5.0